+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -238,6 +238,8 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-block-consolidation-stripe-light-2-background-color': '#fafbfc',
     '--theme-block-consolidation-button-background-color': '#243e62',
     '--theme-block-consolidation-button-background-color-hover': '#fafbfc',
+    '--theme-block-consolidation-button-icon-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-block-consolidation-button-icon-color-hover': '#243e62',
     '--theme-block-consolidation-button-text-color-hover': '#243e62',
     '--theme-block-consolidation-button-border-color': '#fafbfc',
   },
@@ -408,6 +410,9 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-input-right-floating-text-color': 'rgba(94, 96, 102, 0.5)',
     '--theme-input-text-color': '#5e6066',
   },
+  link: {
+    '--theme-link-main-color': 'rgba(41, 111, 208, 1)',
+  },
   loading: {
     '--theme-loading-background-color': '#fafbfc',
     '--theme-loading-no-disk-space-background-color': 'rgba(171, 23, 0, 0.94)',
@@ -425,6 +430,8 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-manual-update-overlay-button-background-color': '#243e62',
     '--theme-manual-update-overlay-button-background-color-hover': '#fafbfc',
     '--theme-manual-update-overlay-button-border-color': '#fafbfc',
+    '--theme-manual-update-overlay-button-icon-color': '#fafbfc',
+    '--theme-manual-update-overlay-button-icon-color-hover': '#243e62',
     '--theme-manual-update-overlay-button-text-color-hover': '#243e62',
     '--theme-manual-update-overlay-text-color': 'rgba(250, 251, 252, 0.7)',
     '--theme-manual-update-overlay-text-highlight-color': '#fafbfc',
@@ -529,6 +536,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-report-issue-connecting-background-color':
       'rgba(250, 251, 252, 0.05)',
     '--theme-report-issue-connecting-text-color': '#fafbfc',
+    '--theme-report-issue-icon-color': '#fafbfc',
     '--theme-report-issue-syncing-background-color': 'rgba(94, 96, 102, 0.05)',
     '--theme-report-issue-syncing-text-color': '#5e6066',
     '--theme-report-issue-syncing-download-logs-text-color': '#5e6066',
@@ -624,6 +632,10 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-staking-delegation-center-gear-icon-fill-color':
       'rgba(94, 96, 102, 0.5)',
     '--theme-staking-delegation-center-gear-icon-fill-color-active': '#5e6066',
+    '--theme-staking-info-learn-more-button-color': '#fafbfc',
+    '--theme-staking-info-learn-more-icon-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-staking-learn-more-button-color': '#fafbfc',
+    '--theme-staking-learn-more-icon-color': 'rgba(250, 251, 252, 0.7)',
   },
   support: {
     '--theme-support-settings-item-color': 'rgba(94, 96, 102, 0.5)',
@@ -637,6 +649,7 @@ const LIGHT_BLUE_THEME_CONFIG = {
   systemError: {
     '--theme-system-error-overlay-attention-icon-color': '#fafbfc',
     '--theme-system-error-overlay-background-color': 'rgba(171, 23, 0, 0.94)',
+    '--theme-system-error-overlay-support-link-icon-color': '#fafbfc',
     '--theme-system-error-overlay-text-color': '#fafbfc',
   },
   tabs: {